+++ forecast/import_actuals.py
@@ -3,7 +3,7 @@ from django.db import connection
 from core.import_csv import get_fk, get_fk_from_field
 from core.models import FinancialYear
 
-from forecast.import_utils import (
+from forecast.utils.import_helpers import (
     CheckFinancialCode,
     UploadFileFormatError,
     sql_for_data_copy,
